+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 110:
 
 * Refactor stream open state variable
+* Refactor websocket stream members
 
 --------------------------------------------------------------------------------